refactor(4d): P3.1 + P3.2 — decompose Gantt timeline + Generate dialog
P3.1: GanttTimeline drops from 551 → 305 lines. The inline ~150-line
task-bar render is now `GanttTaskBar` (memoized on its own props, so a
playback-cursor tick doesn't re-diff every row). Dependency arrows and
the live-drag tooltip also moved out to their own files.

P3.2: GenerateScheduleDialog drops from 567 → 393 lines. The Height-
slice sub-panel and the Advanced disclosure now live in their own
files — `HeightStrategyPanel`, `GenerateAdvancedPanel`. The main dialog
keeps strategy tiles + primary fields + preview.

Neither split changes behaviour — purely readability + memoization
boundaries. Tests stay green (482 viewer + 97 parser + 20 sandbox +
109 cli + every other package). Build clean.
